NetCDF attributes

All attributes in the file, globals are treated as if they belong to variable 'NC_GLOBAL'. Attributes for a single variable may be returned by specifying 'variable' - 'NC_GLOBAL' can stand in to return only those attributes.

nc_atts(x, variable = NULL, ...)

# S3 method for NetCDF
nc_atts(x, variable = NULL, ...)

# S3 method for character
nc_atts(x, variable = NULL, ...)

Arguments

x
filename or handle

variable
optional single name of a variable, or 'NC_GLOBAL'

...
ignored

Value

data frame of attributes

Examples

f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_atts(f)

#> # A tibble: 87 x 3
#>    attribute variable value    
#>        <dbl> <chr>    <list>   
#>  1         0 chlor_a  <chr [1]>
#>  2         1 chlor_a  <chr [1]>
#>  3         2 chlor_a  <chr [1]>
#>  4         3 chlor_a  <dbl [1]>
#>  5         4 chlor_a  <dbl [1]>
#>  6         5 chlor_a  <dbl [1]>
#>  7         6 chlor_a  <chr [1]>
#>  8         7 chlor_a  <dbl [1]>
#>  9         8 chlor_a  <dbl [1]>
#> 10         9 chlor_a  <dbl [1]>
#> # ... with 77 more rows