Top level NetCDF metadata.

This function exists to maintain the open connection while all dimension, variable, and attribute metadata is extracted.

nc_meta(x, ...)

# S3 method for NetCDF
nc_meta(x, ...)

# S3 method for character
nc_meta(x, ...)

Arguments

x
data source address, file name or handle

...
ignored

Details

This function is pretty ambitious, and will send nearly any string to the underlying NetCDF library other than "", which immediately generates an error. This should be robust, but might present fairly obscure error messages from the underlying library.

Examples

f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_meta(f)

#> $dimension
#> # A tibble: 4 x 5
#>      id name          length unlim coord_dim
#>   <int> <chr>          <dbl> <lgl> <lgl>    
#> 1     0 lat             2160 FALSE TRUE     
#> 2     1 lon             4320 FALSE TRUE     
#> 3     2 rgb                3 FALSE FALSE    
#> 4     3 eightbitcolor    256 FALSE FALSE    
#> 
#> $variable
#> # A tibble: 4 x 6
#>      id name    type     ndims natts dim_coord
#>   <int> <chr>   <chr>    <int> <int> <lgl>    
#> 1     0 chlor_a NC_FLOAT     2    12 FALSE    
#> 2     1 lat     NC_FLOAT     1     5 TRUE     
#> 3     2 lon     NC_FLOAT     1     5 TRUE     
#> 4     3 palette NC_UBYTE     2     0 FALSE    
#> 
#> $attribute
#> # A tibble: 87 x 3
#>    attribute     variable value    
#>    <chr>         <chr>    <list>   
#>  1 long_name     chlor_a  <chr [1]>
#>  2 units         chlor_a  <chr [1]>
#>  3 standard_name chlor_a  <chr [1]>
#>  4 _FillValue    chlor_a  <dbl [1]>
#>  5 valid_min     chlor_a  <dbl [1]>
#>  6 valid_max     chlor_a  <dbl [1]>
#>  7 display_scale chlor_a  <chr [1]>
#>  8 display_min   chlor_a  <dbl [1]>
#>  9 display_max   chlor_a  <dbl [1]>
#> 10 scale_factor  chlor_a  <dbl [1]>
#> # ... with 77 more rows
#> 
#> $axis
#> # A tibble: 6 x 3
#>    axis variable dimension
#> * <int> <chr>        <int>
#> 1     1 chlor_a          1
#> 2     2 chlor_a          0
#> 3     3 lat              0
#> 4     4 lon              1
#> 5     5 palette          3
#> 6     6 palette          2
#> 
#> $grid
#> # A tibble: 4 x 2
#>   variable grid 
#>   <chr>    <chr>
#> 1 chlor_a  D1,D0
#> 2 palette  D3,D2
#> 3 lat      D0   
#> 4 lon      D1   
#> 
#> $source
#> # A tibble: 1 x 2
#>   access              source                                                   
#>   <dttm>              <chr>                                                    
#> 1 2018-10-26 20:28:02 /perm_storage/home/mdsumner/Git/ncmeta/inst/extdata/S200…
#> 
#> attr(,"class")
#> [1] "ncmeta"

u <- "https://upwell.pfeg.noaa.gov/erddap/tabledap/FRDCPSTrawlLHHaulCatch"
nc_meta(u)

#> $dimension
#> # A tibble: 2 x 5
#>      id name        length unlim coord_dim
#>   <int> <chr>        <dbl> <lgl> <lgl>    
#> 1     0 maxStrlen64     64 FALSE FALSE    
#> 2     1 s               25 FALSE FALSE    
#> 
#> $variable
#> # A tibble: 18 x 6
#>       id name                     type      ndims natts dim_coord
#>    <int> <chr>                    <chr>     <int> <int> <lgl>    
#>  1     0 s.cruise                 NC_INT        1     4 FALSE    
#>  2     1 s.ship                   NC_CHAR       2     3 FALSE    
#>  3     2 s.haul                   NC_INT        1     4 FALSE    
#>  4     3 s.collection             NC_INT        1     4 FALSE    
#>  5     4 s.latitude               NC_FLOAT      1     8 FALSE    
#>  6     5 s.longitude              NC_FLOAT      1     8 FALSE    
#>  7     6 s.stop_latitude          NC_FLOAT      1     4 FALSE    
#>  8     7 s.stop_longitude         NC_FLOAT      1     4 FALSE    
#>  9     8 s.time                   NC_DOUBLE     1     9 FALSE    
#> 10     9 s.haulback_time          NC_DOUBLE     1     6 FALSE    
#> 11    10 s.surface_temp           NC_FLOAT      1     4 FALSE    
#> 12    11 s.surface_temp_method    NC_CHAR       2     3 FALSE    
#> 13    12 s.ship_spd_through_water NC_FLOAT      1     4 FALSE    
#> 14    13 s.itis_tsn               NC_INT        1     3 FALSE    
#> 15    14 s.scientific_name        NC_CHAR       2     3 FALSE    
#> 16    15 s.subsample_count        NC_INT        1     3 FALSE    
#> 17    16 s.subsample_weight       NC_FLOAT      1     4 FALSE    
#> 18    17 s.remaining_weight       NC_FLOAT      1     4 FALSE    
#> 
#> $attribute
#> # A tibble: 119 x 3
#>    attribute     variable value    
#>    <chr>         <chr>    <list>   
#>  1 actual_range  s.cruise <dbl [2]>
#>  2 description   s.cruise <chr [1]>
#>  3 ioos_category s.cruise <chr [1]>
#>  4 long_name     s.cruise <chr [1]>
#>  5 description   s.ship   <chr [1]>
#>  6 ioos_category s.ship   <chr [1]>
#>  7 long_name     s.ship   <chr [1]>
#>  8 actual_range  s.haul   <dbl [2]>
#>  9 description   s.haul   <chr [1]>
#> 10 ioos_category s.haul   <chr [1]>
#> # ... with 109 more rows
#> 
#> $axis
#> # A tibble: 21 x 3
#>     axis variable         dimension
#>  * <int> <chr>                <int>
#>  1     1 s.cruise                 1
#>  2     2 s.ship                   0
#>  3     3 s.ship                   1
#>  4     4 s.haul                   1
#>  5     5 s.collection             1
#>  6     6 s.latitude               1
#>  7     7 s.longitude              1
#>  8     8 s.stop_latitude          1
#>  9     9 s.stop_longitude         1
#> 10    10 s.time                   1
#> # ... with 11 more rows
#> 
#> $grid
#> # A tibble: 18 x 2
#>    variable                 grid 
#>    <chr>                    <chr>
#>  1 s.ship                   D0,D1
#>  2 s.surface_temp_method    D0,D1
#>  3 s.scientific_name        D0,D1
#>  4 s.cruise                 D1   
#>  5 s.haul                   D1   
#>  6 s.collection             D1   
#>  7 s.latitude               D1   
#>  8 s.longitude              D1   
#>  9 s.stop_latitude          D1   
#> 10 s.stop_longitude         D1   
#> 11 s.time                   D1   
#> 12 s.haulback_time          D1   
#> 13 s.surface_temp           D1   
#> 14 s.ship_spd_through_water D1   
#> 15 s.itis_tsn               D1   
#> 16 s.subsample_count        D1   
#> 17 s.subsample_weight       D1   
#> 18 s.remaining_weight       D1   
#> 
#> $source
#> # A tibble: 1 x 2
#>   access              source                                                   
#>   <dttm>              <chr>                                                    
#> 1 2018-10-26 20:28:07 https://upwell.pfeg.noaa.gov/erddap/tabledap/FRDCPSTrawl…
#> 
#> attr(,"class")
#> [1] "ncmeta"