Top level NetCDF metadata.

This function exists to maintain the open connection while all dimension, variable, and attribute metadata is extracted.

nc_meta(x, ...)

# S3 method for NetCDF
nc_meta(x, ...)

# S3 method for character
nc_meta(x, ...)

Arguments

x
data source address, file name or handle

...
ignored

Details

This function is pretty ambitious, and will send nearly any string to the underlying NetCDF library other than "", which immediately generates an error. This should be robust, but might present fairly obscure error messages from the underlying library.

Examples

f <- system.file("extdata", "S2008001.L3m_DAY_CHL_chlor_a_9km.nc", package = "ncmeta")
nc_meta(f)

#> $dimension
#> # A tibble: 4 x 5
#>      id name          length unlim coord_dim
#>   <int> <chr>          <dbl> <lgl> <lgl>    
#> 1     0 lat             2160 FALSE TRUE     
#> 2     1 lon             4320 FALSE TRUE     
#> 3     2 rgb                3 FALSE FALSE    
#> 4     3 eightbitcolor    256 FALSE FALSE    
#> 
#> $variable
#> # A tibble: 4 x 6
#>      id name    type     ndims natts dim_coord
#>   <int> <chr>   <chr>    <int> <int> <lgl>    
#> 1     0 chlor_a NC_FLOAT     2    12 FALSE    
#> 2     1 lat     NC_FLOAT     1     5 TRUE     
#> 3     2 lon     NC_FLOAT     1     5 TRUE     
#> 4     3 palette NC_UBYTE     2     0 FALSE    
#> 
#> $attribute
#> # A tibble: 87 x 4
#>       id name          variable value    
#>    <int> <chr>         <chr>    <list>   
#>  1     0 long_name     chlor_a  <chr [1]>
#>  2     1 units         chlor_a  <chr [1]>
#>  3     2 standard_name chlor_a  <chr [1]>
#>  4     3 _FillValue    chlor_a  <dbl [1]>
#>  5     4 valid_min     chlor_a  <dbl [1]>
#>  6     5 valid_max     chlor_a  <dbl [1]>
#>  7     6 display_scale chlor_a  <chr [1]>
#>  8     7 display_min   chlor_a  <dbl [1]>
#>  9     8 display_max   chlor_a  <dbl [1]>
#> 10     9 scale_factor  chlor_a  <dbl [1]>
#> # … with 77 more rows
#> 
#> $axis
#> # A tibble: 6 x 3
#>    axis variable dimension
#>   <int> <chr>        <int>
#> 1     1 chlor_a          1
#> 2     2 chlor_a          0
#> 3     3 lat              0
#> 4     4 lon              1
#> 5     5 palette          3
#> 6     6 palette          2
#> 
#> $grid
#> # A tibble: 4 x 4
#>   grid  ndims      variables nvars
#>   <chr> <int> <list<df[,1]>> <int>
#> 1 D1,D0     2        [1 × 1]     1
#> 2 D3,D2     2        [1 × 1]     1
#> 3 D0        1        [1 × 1]     1
#> 4 D1        1        [1 × 1]     1
#> 
#> $source
#> # A tibble: 1 x 2
#>   access              source                                                    
#>   <dttm>              <chr>                                                     
#> 1 2019-10-22 13:11:47 /rdsi/PRIVATE/raad/dev/__tmp_mike/RtmpYDvY2i/temp_libpath…
#> 
#> attr(,"class")
#> [1] "ncmeta"

# \donttest{
if (FALSE) {
u <- "https://upwell.pfeg.noaa.gov/erddap/tabledap/FRDCPSTrawlLHHaulCatch"
nc_meta(u)
}# }